+++ packages/core/src/utils.mjs
@@ -480,7 +480,7 @@ export function mergeI18n(designs, options) {
  * @return {object} result - An object with the merged options and their values
  */
 export function mergeOptions(settings = {}, optionsConfig) {
-  const merged = typeof settings.options === 'undefined' ? {} : { ...settings.option }
+  let merged = {}
   for (const [key, option] of Object.entries(optionsConfig)) {
     if (typeof option === 'object') {
       if (typeof option.pct !== 'undefined') merged[key] = option.pct / 100
@@ -491,6 +491,7 @@ export function mergeOptions(settings = {}, optionsConfig) {
       else if (typeof option.dflt !== 'undefined') merged[key] = option.dflt
     } else merged[key] = option
   }
+  if (typeof settings.options === 'object') merged = { ...merged, ...settings.options }
 
   return merged
 }